+++ CommonplaceBookAppTests/MiniMarkdownParsingTests.swift
@@ -53,6 +53,17 @@ final class MiniMarkdownParsingTests: XCTestCase {
       expectedStructure: "(document (paragraph text (emphasis delimiter text delimiter)))"
     )
   }
+  
+  func testTextWithSingleCharacterEmphasis() {
+    parseText(
+      "Emphasize just the letter *e* in this sentence",
+      expectedStructure: "(document (paragraph text (emphasis delimiter text delimiter) text))"
+    )
+  }
+  
+  func testEmphasisCannotBeEmpty() {
+    parseText("Just ** two stars", expectedStructure: "(document (paragraph text))")
+  }
 
   func testWithBold() {
     parseText(